chore(code): almost always auto (AAA)

+++ lib/ShakeEvents/ShakeEvents.h
@@ -13,7 +13,7 @@ namespace Victoria::Events {
       void loop();
     private:
       uint8_t _inputPin;
-      int _inputState = 0;
+      uint8_t _inputState = 0;
       bool _shaked = false;
       unsigned long _lastFire;
   };
