+++ lib/Timer/Timer.h
@@ -10,21 +10,21 @@ namespace Victoria::Events {
     typedef std::function<void()> TCallback;
     public:
       Timer();
-      unsigned int setTimeout(int timespan, TCallback callback);
-      unsigned int setInterval(int timespan, TCallback callback);
+      unsigned int setTimeout(unsigned long timespan, TCallback callback);
+      unsigned int setInterval(unsigned long timespan, TCallback callback);
       bool clearTimeout(unsigned int id);
       bool clearInterval(unsigned int id);
       void loop();
     private:
       struct Config {
         bool repeat;             // true for setInterval, otherwise setTimeout
-        int timespan;            // delay or interval in milliseconds
+        unsigned long timespan;  // delay or interval in milliseconds
         TCallback callback;      // callback function
         unsigned long timestamp; // timestamp when timer registered
       };
       unsigned int _idSeed = 0;
       std::map<unsigned int, Config> _configs;
-      unsigned int _addConfig(bool repeat, int timespan, TCallback callback);
+      unsigned int _addConfig(bool repeat, unsigned long timespan, TCallback callback);
       bool _removeConfig(unsigned int id);
   };
 } // namespace Victoria::Events
