+++ lib/TimesTrigger/TimesTrigger.h
@@ -8,12 +8,12 @@ namespace Victoria::Events {
   class TimesTrigger {
     typedef std::function<void()> TTimesOutHandler;
     public:
-      TimesTrigger(int times, int resetMillis);
+      TimesTrigger(int times, unsigned long resetMillis);
       TTimesOutHandler onTimesOut;
       void count();
     private:
       int _times;
-      int _resetMillis;
+      unsigned long _resetMillis;
       // state
       int _count = 0;
       unsigned long _lastTime = 0;
